Tp360 fix chrome driver 116 for mac

+++ ui/src/app/modules/dashboard-customer/admin-emails-modal/admin-emails.component.ts
@@ -6,6 +6,7 @@ import { CustomerService } from 'src/app/services/customer.service';
 import { SnackBarService } from 'src/app/services/snack-bar.service';
 import { CustomerAdminEmailService } from "../../../services/customer-admin-email.service";
 import { Observable } from "rxjs";
+import { DialogService } from 'src/app/services/dialog.service';
 
 @Component({
   selector: 'app-admin-emails-modal',
@@ -29,6 +30,7 @@ export class AdminEmailsComponent implements OnInit {
       private customerAdminEmailService: CustomerAdminEmailService,
       private snackBarService: SnackBarService,
       public dialogRef: MatDialogRef<AdminEmailsComponent>,
+      private dialogService: DialogService,
       @Inject(MAT_DIALOG_DATA) public data: any) {
   }
 
@@ -73,18 +75,30 @@ export class AdminEmailsComponent implements OnInit {
 
   deleteExistingEmail(index: number) {
     this.isDataLoading = true;
-    this.customerAdminEmailService.deleteAdminEmail(this.adminEmails[index]).subscribe((res: any) => {
-      if (!res?.error) {
-        this.snackBarService.openSnackBar('Customer administrator email deleted', '');
-        this.adminEmails.splice(index, 1)
-      } else
-        this.snackBarService.openSnackBar(res.error, 'Error while deleting administrator email!');
-      this.isDataLoading = false;
-    }, err => {
-      this.snackBarService.openSnackBar('Error deleting administrator email!');
-      console.error('Error while deleting administrator email', err);
-      this.isDataLoading = false;
+    this.dialogService.confirmDialog({
+      title: 'Confirm action',
+      message: 'Are your sure you want to delete '+this.adminEmails[index] + '?',
+      confirmCaption: 'Delete account',
+      cancelCaption: 'Cancel',
+    }).subscribe((confirmed) => {
+      if(confirmed) {
+        this.customerAdminEmailService.deleteAdminEmail(this.adminEmails[index]).subscribe((res: any) => {
+          if (!res?.error) {
+            this.snackBarService.openSnackBar('Customer administrator email deleted', '');
+            this.adminEmails.splice(index, 1)
+          } else
+            this.snackBarService.openSnackBar(res.error, 'Error while deleting administrator email!');
+          this.isDataLoading = false;
+        }, err => {
+          this.snackBarService.openSnackBar('Error deleting administrator email!');
+          console.error('Error while deleting administrator email', err);
+          this.isDataLoading = false;
+        });
+      } else {
+        this.isDataLoading = false;
+      }
     });
+    
   }
 
   get emailForms() {
